pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "mailsocial"
description = "An email client with social media and instant messaging features."
readme = "README.md"
license = {"text"="GPL-3.0-or-later"}
authors = [
    { name = "James David Clarke", email = "james@jamesdavidclarke.com" }
]
dependencies = [
    "customtkinter",
    "tortoise-orm",
    "pillow",
    "appdirs",
    "python-dotenv",
    "debugpy"
]
requires-python = ">=3.11"

dynamic = ["version"]

[project.scripts]
mailsocial = "src.app:main"

[project.urls]
homepage = "https://github.com/JamesClarke7283/MailSocial"
source = "https://github.com/JamesClarke7283/MailSocial"
tracker = "https://github.com/JamesClarke7283/MailSocial/issues"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-timestamp"

[project.optional-dependencies]
dev = [
    "gitpython",
    "PyGithub",
    "python-gnupg",
    "pyinstaller",
    "black",
    "isort",
    "mypy",
    "bandit"
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
warn_unused_ignores = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_equality = true
follow_imports = "silent"
ignore_missing_imports = true

[tool.bandit]
exclude = ".git,.venv,.mypy_cache,.pytest_cache,.vscode"
skips = ["B101", "B102"]
